requirements.txt
passlib==1.7.4
fastapi==0.66.0
peewee==3.14.0
peewee_async==0.7.0
peewee-migrate==1.1.6
uvicorn==0.14.0
python-jose==3.3.0
python-multipart==0.0.5
starlette==0.14.2
black==20.8b1
fastapi_utils==0.2.1
pyyaml
aiopg==1.3.1
aiofiles==0.7.0
aiohttp==3.7.4.post0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
attrs==21.2.0
h5py==3.3.0
jose==1.0.0
bcrypt==3.2.0
